+++ CHANGELOG.md
@@ -3,6 +3,7 @@ This file documents all updates and releases to the ResearchNow Data Ingest pipe
 
 ## [Unreleased]
 ### Added
+- Publisher data reader to fetch metadata records from smaller projects. Initially this only includes the MET Exhibition Catalogs project.
 - Parser for de Gruyter links for the `sfr-doab-reader`
 ### Fixed
 - Refactored `sfr-doab-reader` tests